split code and tests into their own files
