+++ src/NafIndex.js
@@ -1,5 +1,5 @@
-var options = require('./NafOptions');
-var util = require('./NafUtil');
+var options = require('./options');
+var utils = require('./utils');
 var physics = require('./NafPhysics');
 var NafLogger = require('./NafLogger');
 var Schemas = require('./Schemas');
@@ -11,15 +11,15 @@ naf.app = '';
 naf.room = '';
 naf.clientId = '';
 naf.options = options;
-naf.utils = util;
+naf.utils = utils;
 naf.physics = physics;
 naf.log = new NafLogger();
 naf.schemas = new Schemas();
-naf.version = "0.2.0";
+naf.version = "0.3.0";
 
 var entities = new NetworkEntities();
 var connection = new NetworkConnection(entities);
-naf.connection = naf.c = connection;
-naf.entities = naf.e = entities;
+naf.connection = connection;
+naf.entities = entities;
 
 module.exports = window.NAF = naf;
